provide html file that will generate random card from embedded list
It turns out I want to play this game every time Trump speaks
about the border, and he's sure doing a lot of that these days.
I was drawing out a grid on a piece of paper and looking at the
squares on the screen, and after the nth time I said to myself,
are you a programmer or what?

This is pretty far afield from the original fork now but it
does what I want.