cleanup and add nice terms bingo
The bingo card can be used not just for snark but for nudging the
user to do things to make their day better. Added such a list,
which I will use immediately.

Added option to relabel the free space, added usage example
for nice terms and free space.

Removed dead color code and put the color string in the right
part of the template.