single page bingo that prompts for bingo terms from the user
This displays a bingo card on demand, given a list of phrases
to go on the card, optionally a card title, and a few other
optional settings.

The idea here is to serve this via GitHub Web Pages, thus providing
"bingo as a service".

Unfortunately GitHub does not permit one to pass parameters
in the url for such page, so that bit will have to be done later
and hosted elsewhere.

This is the world's worst html/javascript. It is intended to
work and nothing more than that. Making it nicer might come in
future commits. And then again, it might not.